Moved the FilterResult to the AppCastHandlers folder
