+++ src/NetSparkle/Interfaces/IAppCastFilter.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using NetSparkleUpdater.AppCastHandlers;
 
 namespace NetSparkleUpdater.Interfaces
 {
